refactor(Contributors): replace DataTable with Table in ContributorCompletedPullRequests
