+++ src/app/(user)/tasks/_components/TaskMetricsNumbers.tsx
@@ -1,6 +1,6 @@
 import TaskMetricsCard from '@/app/(user)/tasks/_components/TaskMetricsCard'
 import styles from './TaskMetricsNumbers.module.scss'
-import { Tasks } from '@/app/(user)/contributor/[user_id]/_components/get-tasks'
+import { Tasks } from './get-tasks'
 
 interface TaskMetricsNumbersProps {
   tasks: Tasks
