+++ src/app/(user)/tasks/_components/TaskTypesDistribution.tsx
@@ -1,7 +1,7 @@
 import TaskMetricsCard from '@/app/(user)/tasks/_components/TaskMetricsCard'
 import styles from './TaskTypesDistribution.module.scss'
 import ProgressBar from '@/lib/ui/ProgressBar'
-import { Tasks } from '@/app/(user)/contributor/[user_id]/_components/get-tasks'
+import { Tasks } from './get-tasks'
 
 interface TaskTypesContributionProps {
   tasks: Tasks
